arm64: morello: Unseal the signal handler for PCuABI
On Morello performing exception return (ERET), does not automatically
unseal the capability present in the user signal handler so it should be
unsealed before return to the user mode.

Initialize an unsealing capability for the sentry object type (RB in
Morello) derived from the root capability. This may be used to unseal the
signal handler.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>