drivers/android/binder_alloc: Represent user addresses as integers
A couple of structs internal to binder_alloc represent addresses to
user pages as void __user *. In PCuABI, this is a bad idea, as this
type should only be used to represent "real" user pointers, not
merely addresses. If a user pointer is set to an integer address
(not a valid pointer) and subsequently cast back to an integer, the
resulting value is zero.

Fix this by representing such addresses as unsigned long, as per the
convention used in mm. This also greatly reduces the need for casts
when working with them.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>