usb: core: Fix __capability annotations in double pointers
CHERI compilers mandatorily need __capability annotations after the
pointer[*] in case of double pointers.

In order to minimise the need to disambiguate double user pointers,
user pointers are now passed around as straightforward void __user *
instead of double user pointers. This also allows us to get rid of a
number of casts. Note that the actual type that processcompl() takes
is a double user pointer (as per the cast when calling put_user()).

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>