usb: Support 64-bit pointers in compat
In compat64, pointers are 64-bit, and cannot be represented as u32;
the appropriate type to represent compat pointers is compat_uptr_t.
Adjust processcompl_compat() accordingly. Since the definition of
USBDEVFS_REAPURB{,NDELAY} is based on the size of void * from a
userspace perspective, USBDEVFS_REAPURB{,NDELAY}32 also need to be
adjusted.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>