arm64: compat64: Use native handler for getrusage syscall
A bug was found in the getrusage syscall where the compat64 version used
a different sized struct to pass arguments. This resulted in some
programs such as time in bash printing wrong values for process usage
times.

This is being fixed by using the native syscall which has the same
implementation since the struct members contain no pointers.

Signed-off-by: Sherwin da Cruz <sherwin.dacruz@arm.com>