arm64: morello: Disable SHA3 detection
We have already taken care of such problematic features for v8.3+,
see commit "arm64: morello: Disable v8.3+ AS-dependent features".

We now need to do the same for SHA3. It is the only v8.2 optional
feature that requires AS support and it is not implemented on
Morello, so we can use the same trick as for v8.3+ and disable
AS_HAS_ARMV8_2 if Morello is selected. We also need to disable
AS_HAS_SHA3 as it implicitly depends on AS_HAS_ARMV8_2 to compile
functions using SHA3 instructions.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>