epoll: Relax epitem size check for CHERI systems
For performance reasons eventpoll_init() currently checks that the size
of struct epitem is kept under 128 bytes. It avoids making this check
for 128-bit and larger systems, as on those systems struct epitem would
be above 128 bytes in size.

On a CONFIG_CHERI_PURECAP_UABI kernel on a CHERI system, 'long' and
'void *' remain 64-bit but 'void __user *' is 128-bit which causes
struct epitem to go over 128 bytes in size. Relax the check for such
systems as well.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>