epoll: Use user pointer types for epoll_event.data
The 'data' member of struct epoll_event may contain a user pointer. On
some architectures (for example CHERI) a user pointer is a 129-bit
capability, so the __u64 type is not big enough to hold it. Use the
__kernel_uintptr_t type instead, which is big enough on the affected
architectures while remaining 64-bit on others.

In addition, special copy routines need to be used when copying user
pointers to and from userspace. Use these for epoll_event.data.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>