kernel: signal: Update pidfd_send_signal() syscall for PCuABI
pidfd_send_signal syscall is used to signal and pass siginfo_t type data
to a process identified by a pidfd. PCuABI specification requires that the
data should be transmitted along with the capability pointer tags in case
of signal delivered to the same process. In case of different process no
tag details should be copied and any tag present in siginfo_t should be
treated as an -EINVAL type error.

Modify function copy_siginfo_from_user_any() to have a bool argument and is
obtained by comparing received pidfd with the current process. However the
siginfo_t tag validation checks for non current process is marked as a TODO
for now.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>