bpf: Use proper typecast for capability type
With the introduction of capabilities, the PCuABI expects a
capability when dealing with the user pointers.

Address the compilation warning below triggered by otherwise
implicit conversion that might lead to unexpected behaviour when
operating on capabilities.

make[1]: linux/kernel/bpf/helpers.c:
warning: the following conversion will result in a CToPtr operation;
the behaviour of CToPtr can  be confusing since using CToPtr on an
untagged capability will give 0 instead of the  integer value and should
therefore be explicitly annotated [-Wcheri-pointer-conversion]
  ret = access_process_vm(tsk, (unsigned long)user_ptr, dst,size, 0);

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>