linux/cheri.h: Introduce CHERI helpers
For architectures implementing CHERI capabilities, introduce a new
cheri.h header that defines a few helpers and macros, in addition to
what is available in the compiler-provided <cheriintrin.h>.

This header is only meant for code manipulating CHERI capabilities
explicitly and expands to nothing if __CHERI__ is not defined.

CONFIG_HAVE_ARCH_CHERI_H is also added to allow architectures to
override some of the definitions by providing their own
asm/cheri.h.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>