linux/user_ptr.h: Introduce uaccess-related helpers
Introduce new helpers that help interact with user pointers for the
purpose of accessing user memory, in a PCuABI-friendly way:

* make_user_ptr_for_<perms>_uaccess() creates a user capability with
  the requested bounds and permissions, strictly to perform uaccess.

* check_user_ptr_<perms>() checks that the capability is valid and
  its bounds and permissions are appropriate for the requested
  access. This is needed if (and only if) user memory is accessed
  via kernel mappings, and capability metadata therefore needs to be
  checked upfront in PCuABI. When normal uaccess is performed,
  capability metadata is checked directly by the hardware; these
  helpers can be seen as the counterparts of such checks for
  "indirect uaccess" via kernel mappings.

In the PCuABI case, these functions are implemented on top of the
new cheri_* helpers in linux/cheri.h. The implementation is trivial
in the !PCuABI case (the specified length and permissions are
ignored).

make_user_ptr_for_<perms>_uaccess() replaces uaddr_to_user_ptr_safe()
to create user pointers to perform uaccess. The latter is still
used (for the time being) to provide new user pointers to userspace.

The comments and documentation are updated accordingly.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>