bpf: copy_{to,from}_user_with_ptr helpers
In PCuABI, when copying a block of memory to/from userspace containing
capabilities/pointers, the copy_{to,from}_user_with_ptr variants need
to be used to ensure capabilities are preserved in full.

Introduce bpf specific helper methods/macros to support this.

Since compat64 syscalls do not contain capabilities, the _with_ptr
variants are not required in the in_compat64_syscall() case.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>