bpf: use user pointer types in uAPI structs
union bpf_attr and structs bpf_xyz_info are used to pass config to the
kernel or return info to userspace via the bpf syscall.

In order to avoid needing a separate 32-bit compat handler for the bpf
syscall, __user pointers inside these union/structs are stored as __u64
or __aligned_u64.

In PCuABI, a user pointer is a 129-bit capability so __aligned_u64 type
is not big enough to hold it. Use type __kernel_aligned_uintptr_t
instead, which is big enough on the affected architectures while
remaining __aligned_64 on others.

Use copy_{to,from}_user_with_ptr variants where blocks of memory
containing pointers are being copied to preserve capabilities.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>