Documentation: Add CHERI / PCuABI document
PCuABI itself is defined by the PCuABI specification. However, the
specification only documents the ABI itself and not internal kernel
implementation aspects. To that effect, create a document under a
new cheri/ subfolder, as well as an index file with some information
about CHERI support in general.

Now that we have a generic PCuABI document, link to it from the
related documents, and remove a now-redundant section from the user
pointer doc. All CHERI / PCuABI-related documents are now reachable
from Documentation/cheri/index.rst.

The PCuABI porting guide was initially added to the root of
Documentation/ for lack of relevant subfolder, we can now move it to
a more appropriate home.

Reviewed-by: Vincenzo Frascino <vincenzo.frascino@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>