ptrace: fix PTRACE_POKEDATA write size
The PTRACE_POKEDATA request writes a word of data to the tracee's
memory. In PCuABI the size of the write remains 8 bytes. Currently the
kernel erroneously writes 16 bytes, thereby overwriting 8 bytes of
unrelated memory. Fix this by restoring the type of the data argument of
generic_ptrace_pokedata() to unsigned long.

Fixes: ("kernel/ptrace: Modify ptrace syscall to accept capability arguments")
Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>