arm64: morello: Initialise user capabilities from cheri_user_root_*
Now that we have introduced the generic cheri_user_root_* userspace
root capabilities, we can make use of them to derive the initial
user capabilities (process startup and signal delivery). Note that
DDC will be taken care of separately, as it is not saved in
struct pt_regs and is not initialised in morello_thread_start()
either.

For purecap (PCuABI) processes, we derive the capabilities from
the "regular" root capability; their bounds and permissions will
later be reduced further as per the PCuABI specification.

For hybrid processes (i.e. plain AArch64 ABI + capability support),
the capabilities are derived from the special root capability with all
permissions, cheri_user_root_allperms_cap. That's because we are not
attempting to constrain the capabilities provided to hybrid
processes in any way, and there is no mechanism to provide
special permissions (Seal/Unseal/CompartmentID) separately in hybrid.

Since the purecap/hybrid initialisation is now clearly separated, we
also use that opportunity to leave CSP zero-initialised in hybrid;
a valid CSP should only be provided in purecap.

To avoid some ungraceful #ifdef'ing in
morello_setup_signal_return(), morello_sentry_unsealcap is now
always defined, as the overhead is insignificant (a 16-byte global
and a few instructions on startup).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>