arm64: morello: Initialise user DDC from cheri_user_root_*
Just like PCC/CSP, we can now derive the initial user DDC from the
appropriate cheri_user_root_* userspace root capability. We still
provide it in purecap to be consistent with the transitional PCuABI
specification, but it will eventually be set to the null capability.

Checking whether we are in compat or not in assembly would be
tedious, so a C wrapper is used to do the purecap/hybrid selection.
morello_thread_init_user() always targets current, so to keep things
simple and use the argumentless is_pure_task() helper, we remove its
task_struct * argument. This is in line with other helpers called
from arch_setup_new_exec().

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>