kselftests/arm64: morello: Process the dynamic relocations
Gcc emits only dynamic relocations and not cap relocations like Clang
toolchain. Hence the existing kselftests compiled in static and standalone
mode need to process the dynamic relocations and fill the entries to
relevant capability details.

The code changes are similar to the start up code of the recently released
Arm GNU Toolchain for Morello.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>