arm64: compat: Only define compat_mode_t as u16 in compat32
mode_t is u16 only in 32-bit. In compat64, compat_mode_t should be
the same as the 64-bit mode_t, that is u32. This is the type
compat_mode_t is defined to in <asm-generic/compat.h> by default, so
simply add an #ifdef CONFIG_COMPAT32.

This fixes the definition of struct compat_ipc64_perm in compat64.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>