io_uring/tctx: Copy only the offset field back to user
Upon successful return of the io_uring_register system call, the offset
field will contain the value of the registered file descriptor to be
used for future io_uring_enter system calls. The rest of the struct
doesn't need to be copied back to userspace, so restrict the
copy_to_user call only to the offset field.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>