io_uring: Use user pointer type in the uAPI structs
Some members of the io_uring uAPI structs may contain user pointers. In
the PCuABI, a user pointer is a 129-bit capability, so the __u64 type is
not big enough to hold it. Use the __kernel_uintptr_t type instead,
which is big enough on the affected architectures while remaining 64-bit
on others.

The user_data field must be passed unchanged from the submission queue
to the completion queue. As it is standard practice to store a pointer
in user_data, expand the field to __kernel_uintptr_t. However, the
kernel doesn't dereference the user_data, so don't convert it in the
compat case.

In addition, for the io_uring structs containing user pointers, use the
special copy routines when copying user pointers from/to userspace.

To ensure that a full capability comparison is performed in the cases
where two user data values are compared, a new helper is introduced.

In the case of operation IORING_OP_POLL_REMOVE, if
IORING_POLL_UPDATE_USER_DATA is set in the SQE len field, then the
request will update the user_data of an existing poll request based on
the value passed in the addr2 field, instead of the off field. This is
required because the off field is not large enough to fit a user_data
value.

Note that the structs io_uring_sqe and io_uring_cqe are doubled in size
in PCuABI. The setup flags IORING_SETUP_SQE128 and IORING_SETUP_CQE32
used to double the sizes of the two structs up to 128 bytes and 32 bytes
respectively. In PCuABI, the two flags are still used to double the
sizes of the two structs, but, as they increased in size, they increase
up to 256 bytes and 64 bytes, respectively.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>