net: bridge: Support 64-bit pointers in compat
The br_dev_read_uargs function copies a list of arguments
from a user capability. When operating in compat64, pointers
are 64-bit long and cannot be represented as unsigned int.
In this scenario the function copies a list of compat pointers,
therefore the type of the receiver variable, cargs, must be
updated to compat_uptr_t from unsigned int.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>