coredump: Explicitly create user pointer
dump_vma_snapshot() is directly reading memory from a user mapping.
The access is entirely driven by the kernel, not by userspace, so
a user pointer has to be created manually from the computed address.
Using uaddr_to_user_ptr_safe() is the appropriate way to create a
user pointer in this case.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>