audit: Explicitly create user pointer
audit_log_execve_info() is directly reading memory from a user mapping.
The access is entirely driven by the kernel, not by userspace, so
a user pointer has to be created manually from the address of the
arguments region (mm->arg_start). Using uaddr_to_user_ptr_safe() is
the appropriate way to create a user pointer in this case.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>