kselftests/arm64: morello: clone: Remove loop check
Some of the clone tests use loop check on status flag to determine if the
child exited normally. However, this may cause infinite looping if the
data handshake does not happen properly between parent and child due to
compiler optimization as observed with Morello Gcc.

Although this can be fixed by setting the flag to volatile but here we
replace the busy loop with the waitpid() syscall to wait for the child to
exit before parent is executed. This approach allows us to get rid of
item status and result from struct test_fixture.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>