aio: Implement compat handling for the io_event struct
As the io_event struct might contain pointers as members, introduce a
compat version of the struct. Also, implement functions that convert
the compat version to the native version of the struct.

A subsequent patch is going to change the io_event struct to enable it
to support new architectures. On such architectures, the current struct
layout still needs to be supported for compat tasks.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>