fs/binfmt_elf: Create appropriate user pointer for uaccess
The ELF loader needs to zero out part of the BSS explicitly. It also
needs to read the program arguments directly from user memory when
dumping a core file. In both cases we only have the address of the
memory mapping / arguments to access, and need to create a valid
user pointer to perform the uaccess.

uaddr_to_user_ptr_safe() should no longer be used for that purpose.
Instead, we use make_user_ptr_for_{read,write}_uaccess() to create a
user pointer with appropriate bounds and permissions (in PCuABI).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>